requirements.txt
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.3
fonttools==4.56.0
fqdn==1.5.1
fsspec==2026.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.8
lark==1.3.1
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mistune==3.1.4
ml_dtypes==0.5.4
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.3.5
onnx==1.20.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.5
pillow==11.1.0
pipreqs==0.4.13
platformdirs==4.5.0
polars==1.37.1
polars-runtime-32==1.37.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.4
psutil==7.1.3
pure_eval==0.2.3
pycparser==2.23
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.1
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.28.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.10.16
tinycss2==1.4.0
torch==2.9.1+cu126
torchvision==0.24.1+cu126
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.1
ultralytics==8.4.5
ultralytics-thop==2.0.18
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
yarg==0.1.10
